legacy(py)/skyth-old/requirements.txt
# Web Framework & Async
quart==0.19.4
quart-cors==0.7.0
hypercorn==0.16.0
flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
quart-rate-limiter==0.9.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# AI & Machine Learning
google-genai==0.2.1
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.0
click==8.1.8
requests==2.31.0
PyJWT==2.8.0
beautifulsoup4==4.12.3
weasyprint==60.2
pillow==10.2.0

# Tools & APIs
yfinance==0.2.36
pandas==2.2.0
youtube-transcript-api==0.6.2
spotipy==2.23.0
wikipedia-api==0.6.0
selenium==4.17.2
duckduckgo-search==4.4.1
ddgs